requirements.txt
oauthlib==3.2.0
python-dotenv==1.0.0
python_graphql_client==0.4.3
Requests==2.31.0
requests_oauthlib==1.3.1
